backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0

# Supabase
supabase>=2.3.0

# AI and Audio
anthropic>=0.18.0
elevenlabs>=1.1.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.12.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiofiles>=23.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Logging and monitoring
structlog>=24.1.0
